RS-XX: Update release pipeline
In the last delivery the amount of referenced workflows exceeded the
GitHub actions limit of 20. In order to prevent a failure of the delivery
this update aims to build all IPF containers through a new general
action, that will be build through a matrix strategy. This strategy
allows for up to 256 jobs of the same pattern (here we are using 22).