RS-595: Update S2 metadata extraction to include all fields for dissemination
Please note, that as the V2 does not include the Inventory_Metadata.xml for S2 L0 products
yet, the extraction will be handled differently for those product types. Instead of
extracting most of the information from the Inventory_Metadata.xml we get what we can
from the other metadata file and have to leave other fields empty.